requirements.txt
mypy==0.501
nose2==0.6.5
pyyaml==5.1b3